Organized tic tac toe game into modular files.
