+++ src/routes/playground/tic-tac-toe.svelte
@@ -1,0 +1,17 @@
+<script lang="ts">
+    import TicTacToeGame from "$lib/playground/tic-tac-toe/Game.svelte";
+</script>
+
+<div class="window">
+    <TicTacToeGame />
+</div>
+
+<style>
+    .window {
+        width: 100vw;
+        height: 100vh;
+        display: flex;
+        justify-content: center;
+        align-items: center;
+    }
+</style>
